+++ TabletopTweaks/NewContent/Feats/ExtraReservoir.cs
@@ -14,7 +14,7 @@ namespace TabletopTweaks.NewContent.Feats {
 
             var ExtraReservoir = FeatTools.CreateExtraResourceFeat("ExtraReservoir", ArcanistArcaneReservoirResource, 3, bp => {
                 bp.SetName("Extra Reservoir");
-                bp.SetDescription("You gain three more points in your arcane reservoir, and the maximum number of points in your arcane reservoir " +
+                bp.SetDescription("You gain 3 more points in your arcane reservoir, and the maximum number of points in your arcane reservoir " +
                     "increases by that amount.\nYou can take this feat multiple times. Its effects stack.");
                 bp.AddPrerequisiteFeature(ArcanistArcaneReservoirFeature, GroupType.Any);
                 bp.AddPrerequisiteFeature(ArcaneEnforcerArcaneReservoirFeature, GroupType.Any);